monitoring-jobs/requirements.txt
mysql-connector==2.2.9
prometheus_client==0.15.0